fixed 1.5 -> 1.4 build errors
git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@30077 0d108963-8a47-0410-809b-ebd9fb3b3407

+++ upl/com.energyict.mdc.upl.impl/protocolimpl/src/main/java/com/energyict/protocolimpl/modbus/flonidan/uniflo1200/UNIFLO1200.java
@@ -104,7 +104,7 @@ public class UNIFLO1200 extends Modbus {
 			getRegisterFactory().findRegister(UNIFLO1200RegisterFactory.REG_LOGIN).getWriteMultipleRegisters(b);
 		}
 		
-		setSecLvl((Integer)getRegisterFactory().findRegister(UNIFLO1200RegisterFactory.REG_ACTUAL_SECLEVEL).value());		
+		setSecLvl(((Integer)getRegisterFactory().findRegister(UNIFLO1200RegisterFactory.REG_ACTUAL_SECLEVEL).value()).intValue());		
 
 		if (getInfoTypeSecurityLevel() != getSecLvl())
 			throw new InvalidPropertyException("SecurityLevel mismatch [" + getInfoTypeSecurityLevel() + " != " + getSecLvl() + "]: Reason may be wrong password or hardware lock.");
